Promote profile_description and periodic_ts to core timeseries
